Updated project to work with XBuilder

+++ src/C/render/palette.c
@@ -287,9 +287,9 @@ X3D_ColorIndex x3d_colorindex_from_rgb(uint8 r, uint8 g, uint8 b) {
     X3D_ColorIndex min_index = 0;
     
     for(i = 0; i < 256; ++i) {        
-        int16 dist = abs((int16)r - quake_color_palette[i][0]) +
-            abs((int16)g - quake_color_palette[i][1]) +
-            abs((int16)b - quake_color_palette[i][2]);
+        int16 dist = abs((int16)r - (int16)quake_color_palette[i][0]) +
+            abs((int16)g - (int16)quake_color_palette[i][1]) +
+            abs((int16)b - (int16)quake_color_palette[i][2]);
             
         if(dist < min_dist) {
             min_dist = dist;
